Refactor the remaining types used by CKB RPC

+++ core/src/main/java/org/nervos/ckb/type/Cycles.java
@@ -1,7 +1,5 @@
 package org.nervos.ckb.type;
 
-/** Copyright © 2019 Nervos Foundation. All rights reserved. */
 public class Cycles {
-
-  public int cycles;
+  public long cycles;
 }
